Changed enum Dpl representation

+++ src/interrupts/x86_64/mod.rs
@@ -258,7 +258,9 @@ impl TrapGate {
 
     /// Get Descriptor Privilege Level.
     pub fn dpl(&self) -> Dpl {
-        unsafe { ::core::mem::transmute(self.flags & 0b01100000_00000000) }
+        unsafe {
+            ::core::mem::transmute((self.flags & 0b01100000_00000000) as u8)
+        }
     }
 
     /// Set Descriptor Privilege Level.
